Clarified backup position.  It's not due to lack of functionality but rather intentional feature differentiation.  Included link to xtrabackup (very popular).
The query cache is now disabled by default in mysql 5.6.

The innodb buffer pool is automatically partitioned when larger than 1GB, thus providing cpu scalability.  The myisam key cache is not, and is still one single mutex protecting loading/unloading, and does not scale.

The table_cache is partitioned in mysql 5.6.  The main lock that was painful was lock_open mutex, which was broken down into smaller locks starting with work in 5.5 with meta data locking introduced.